Add readme and fix frontend
